Modified latest file lookup.
